backend/requirements.txt
fastapi
uvicorn[standard]
gensim
faiss-cpu
pandas
scikit-learn
pydantic
nltk